cleaning up inheritance of persistent/direct interfaces
