+++ pyomo/solvers/plugins/solvers/gurobi_direct.py
@@ -17,6 +17,7 @@ from pyomo.util.plugin import alias
 from pyomo.core.kernel.numvalue import is_fixed
 from pyomo.repn import generate_canonical_repn, LinearCanonicalRepn, canonical_degree
 from pyomo.solvers.plugins.solvers.direct_solver import DirectSolver
+from pyomo.solvers.plugins.solvers.direct_or_persistent_solver import DirectOrPersistentSolver
 from pyomo.core.kernel.numvalue import value
 import pyomo.core.kernel
 from pyomo.core.kernel.component_set import ComponentSet
@@ -80,9 +81,6 @@ class GurobiDirect(DirectSolver):
         self._capabilities.integer = True
         self._capabilities.sos1 = True
         self._capabilities.sos2 = True
-
-    def _presolve(self, *args, **kwds):
-        DirectSolver._presolve(self, *args, **kwds)
 
     def _apply_solver(self):
         if self._tee:
@@ -196,7 +194,7 @@ class GurobiDirect(DirectSolver):
 
     def _compile_instance(self, model, kwds={}):
         self._range_constraints = set()
-        DirectSolver._compile_instance(self, model, kwds)
+        DirectOrPersistentSolver._compile_instance(self, model, kwds)
         try:
             self._solver_model = self._gurobipy.Model()
         except Exception:
@@ -567,7 +565,7 @@ class GurobiDirect(DirectSolver):
         # manager, created populated *directly* by this plugin.
         pyutilib.services.TempfileManager.pop(remove=not self._keepfiles)
 
-        return DirectSolver._postsolve(self)
+        return DirectOrPersistentSolver._postsolve(self)
 
     def warm_start_capable(self):
         return True
